Tests/LibWeb/Text/expected/element-get-bounding-client-rect-css-transform.txt

       {"x":68,"y":118,"width":100,"height":100,"top":118,"right":168,"bottom":218,"left":68}
